Fix creation of bank clearing account
Since issue6003, setting the parent account set some fields but the scenarios
set the values using the **kwargs option of proteus. In this case the order
used to set the fields is undefined until PEP468, so the parent could be set
after other fields and create a misconfigured account which will make the
scenario fails later when creating the clearing move.